+++ app/src/main/java/com/orange/ods/app/ui/components/ComponentsScreen.kt
@@ -31,7 +31,7 @@ import com.orange.ods.app.R
 import com.orange.ods.app.ui.LocalMainTopAppBarManager
 import com.orange.ods.app.ui.utilities.DrawableManager
 import com.orange.ods.compose.component.card.OdsSmallCard
-import com.orange.ods.utilities.extension.orElse
+import com.orange.ods.extension.orElse
 
 @Composable
 fun ComponentsScreen(onComponentClick: (Long) -> Unit) {