+++ lib/src/main/java/com/orange/ods/compose/component/card/OdsVerticalHeaderFirstCard.kt
@@ -38,7 +38,7 @@ import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.text.OdsTextBody1
 import com.orange.ods.compose.text.OdsTextH6
 import com.orange.ods.compose.text.OdsTextSubtitle2
-import com.orange.ods.utilities.extension.orElse
+import com.orange.ods.extension.orElse
 
 /**
  * <a href="https://system.design.orange.com/0c1af118d/p/272739-cards/b/991690" target="_blank">ODS Card</a>.